create TextButton, replace most pressables with text inside with TextButton component, move multiple components into new 'common' folder
