+++ components/game_components/GuessWarningPopup.tsx
@@ -4,7 +4,7 @@ import { StyleSheet } from "react-native";
 import { Modal } from "react-native";
 import Word from "./Word";
 import { WordState } from "./Word";
-import GenericPopup from "../GenericPopup";
+import GenericPopup from "../common/GenericPopup";
 
 
 interface GuessWarningPopupProps {
